refactor: 커스텀 훅에 작성했던 계산관련 로직들 utils로 분리
